Split oe-selftests by target machines
To allow for more targeted testing
